refactor(ui): flatten `cron-workflows/` dir
- same as previous commits
- waited for a bit before tackling this dir as it has more components than others
  - but still only around ~10, so not _that_ many
