--- ui/src/app/cron-workflows/components/cron-workflow-details/cron-workflow-details.scss
@@ -1,14 +1,0 @@
-@import 'node_modules/argo-ui/src/styles/config';
-
-.workflows-cron-list .workflows-list {
-  &__status {
-    max-width: 80px;
-    display: flex;
-    align-items: center;
-
-    /* checkboxes are not visible and are unused on this page */
-    &--checkbox {
-      appearance: none;
-    }
-  }
-}
