+++ ui/src/app/cron-workflows/cron-workflow-details.tsx
@@ -2,21 +2,22 @@ import {NotificationType, Page, SlidingPanel} from 'argo-ui';
 import * as React from 'react';
 import {useContext, useEffect, useState} from 'react';
 import {RouteComponentProps} from 'react-router';
-import * as models from '../../../../models';
-import {CronWorkflow, Link, Workflow} from '../../../../models';
-import {uiUrl} from '../../../shared/base';
-import {ErrorNotice} from '../../../shared/components/error-notice';
-import {openLinkWithKey} from '../../../shared/components/links';
-import {Loading} from '../../../shared/components/loading';
-import {useCollectEvent} from '../../../shared/components/use-collect-event';
-import {ZeroState} from '../../../shared/components/zero-state';
-import {Context} from '../../../shared/context';
-import {historyUrl} from '../../../shared/history';
-import {services} from '../../../shared/services';
-import {useQueryParams} from '../../../shared/use-query-params';
-import {WidgetGallery} from '../../../widgets/widget-gallery';
-import {WorkflowsRow} from '../../../workflows/components/workflows-row/workflows-row';
-import {CronWorkflowEditor} from '../cron-workflow-editor';
+
+import * as models from '../../models';
+import {CronWorkflow, Link, Workflow} from '../../models';
+import {uiUrl} from '../shared/base';
+import {ErrorNotice} from '../shared/components/error-notice';
+import {openLinkWithKey} from '../shared/components/links';
+import {Loading} from '../shared/components/loading';
+import {useCollectEvent} from '../shared/use-collect-event';
+import {ZeroState} from '../shared/components/zero-state';
+import {Context} from '../shared/context';
+import {historyUrl} from '../shared/history';
+import {services} from '../shared/services';
+import {useQueryParams} from '../shared/use-query-params';
+import {WidgetGallery} from '../widgets/widget-gallery';
+import {WorkflowsRow} from '../workflows/components/workflows-row/workflows-row';
+import {CronWorkflowEditor} from './cron-workflow-editor';
 
 import '../../../workflows/components/workflow-details/workflow-details.scss';
 import './cron-workflow-details.scss';
