+++ ui/src/app/cron-workflows/cron-workflow-link.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
-import {uiUrl} from '../../shared/base';
-import {LinkButton} from '../../shared/components/link-button';
+
+import {uiUrl} from '../shared/base';
+import {LinkButton} from '../shared/components/link-button';
 
 export const CronWorkflowLink = (props: {namespace: string; name: string}) => (
     <LinkButton to={uiUrl('cron-workflows/' + props.namespace + '/' + props.name)}>
