+++ ui/src/app/cron-workflows/cron-workflow-list.tsx
@@ -2,26 +2,27 @@ import {Page, SlidingPanel, Ticker} from 'argo-ui';
 import * as React from 'react';
 import {useContext, useEffect, useState} from 'react';
 import {Link, RouteComponentProps} from 'react-router-dom';
-import {CronWorkflow, CronWorkflowSpec} from '../../../../models';
-import {ANNOTATION_DESCRIPTION, ANNOTATION_TITLE} from '../../../shared/annotations';
-import {uiUrl} from '../../../shared/base';
-import {ErrorNotice} from '../../../shared/components/error-notice';
-import {ExampleManifests} from '../../../shared/components/example-manifests';
-import {InfoIcon} from '../../../shared/components/fa-icons';
-import {Loading} from '../../../shared/components/loading';
-import {Timestamp} from '../../../shared/components/timestamp';
-import {useCollectEvent} from '../../../shared/components/use-collect-event';
-import {ZeroState} from '../../../shared/components/zero-state';
-import {Context} from '../../../shared/context';
-import {getNextScheduledTime} from '../../../shared/cron';
-import {Footnote} from '../../../shared/footnote';
-import {historyUrl} from '../../../shared/history';
-import {services} from '../../../shared/services';
-import {useQueryParams} from '../../../shared/use-query-params';
-import {Utils} from '../../../shared/utils';
-import {CronWorkflowCreator} from '../cron-workflow-creator';
-import {CronWorkflowFilters} from '../cron-workflow-filters/cron-workflow-filters';
-import {PrettySchedule} from '../pretty-schedule';
+
+import {CronWorkflow, CronWorkflowSpec} from '../../models';
+import {uiUrl} from '../shared/base';
+import {ANNOTATION_DESCRIPTION, ANNOTATION_TITLE} from '../shared/annotations';
+import {ErrorNotice} from '../shared/components/error-notice';
+import {ExampleManifests} from '../shared/components/example-manifests';
+import {InfoIcon} from '../shared/components/fa-icons';
+import {Loading} from '../shared/components/loading';
+import {Timestamp} from '../shared/components/timestamp';
+import {useCollectEvent} from '../shared/use-collect-event';
+import {ZeroState} from '../shared/components/zero-state';
+import {Context} from '../shared/context';
+import {getNextScheduledTime} from '../shared/cron';
+import {Footnote} from '../shared/footnote';
+import {historyUrl} from '../shared/history';
+import {services} from '../shared/services';
+import {useQueryParams} from '../shared/use-query-params';
+import {Utils} from '../shared/utils';
+import {CronWorkflowCreator} from './cron-workflow-creator';
+import {CronWorkflowFilters} from './cron-workflow-filters';
+import {PrettySchedule} from './pretty-schedule';
 
 import './cron-workflow-list.scss';
