+++ ui/src/app/cron-workflows/cron-workflow-spec-editior.tsx
@@ -1,8 +1,9 @@
 import {Checkbox, Select} from 'argo-ui';
 import * as React from 'react';
-import {ConcurrencyPolicy, CronWorkflowSpec} from '../../../models';
-import {NumberInput} from '../../shared/components/number-input';
-import {TextInput} from '../../shared/components/text-input';
+
+import {ConcurrencyPolicy, CronWorkflowSpec} from '../../models';
+import {NumberInput} from '../shared/components/number-input';
+import {TextInput} from '../shared/components/text-input';
 import {ScheduleValidator} from './schedule-validator';
 
 export function CronWorkflowSpecEditor({onChange, spec}: {spec: CronWorkflowSpec; onChange: (spec: CronWorkflowSpec) => void}) {
